requirements.txt
# GEOEventFusion — Core Runtime Dependencies
# Install via: pip install -e .
# Or directly: pip install -r requirements.txt

# HTTP clients
requests>=2.31.0
httpx>=0.27.0

# Feed ingestion and full-text extraction
feedparser>=6.0.11
trafilatura>=1.9.0
newspaper3k>=0.2.8

# Graph analysis
networkx>=3.2.1

# LLM backends
anthropic>=0.34.0
ollama>=0.3.0

# Configuration and environment
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
PyYAML>=6.0.1

# Numerical and data
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0

# Visualization
matplotlib>=3.8.0
folium>=0.17.0

# Text utilities
Levenshtein>=0.25.0
python-dateutil>=2.9.0

# Progress
tqdm>=4.66.0
